feat: add the CreditScore and CreditScoreHistory models to entities layer
This facilitates the definition of a CreditScore and CreditScoreHistory object within the context of the Ava application.

This will be used to access the necessary data for the credit score related widgets. This places the models in the `/entities/credit_score` directory to adhere to the architecture principles of Feature-Sliced Design and MVVM.

<!-- ps-id: eb488906-dc99-418a-896b-1da471b99cf0 -->