feat: add CreditScoreHeroCard view
This facilitates the displaying of the credit score with an animated progress indicator, the dates of the last and next reporting cycles, and the change in credit since the last reporting.

This will be displayed on the Home page. This adheres to the architecture principles of Feature-Sliced Design and MVVM. This patch also removes the custom TextTheme until we can get access to the actual font-files.

<!-- ps-id: 07e4b21c-a70a-4344-a0d7-f917916e3dac -->